🔨 Refactor: Banner 컴포넌트 위치 변경
기존에는 SignUp 폴더에 있었음. 하지만 SignUp 뿐만 아니라 Login에서도 쓰기 때문에 SignUp 폴더 밖으로 이동시킴
