+++ src/Component/Banner.tsx
